file has been componentised
