--- src/.env
@@ -1,2 +1,0 @@
-REACT_APP_API_BASE_URL=https://bus-app-api-kl95.onrender.com
-REACT_APP_MAPBOX_ACCESS_TOKEN=pk.eyJ1IjoiemFjYm1yMjIiLCJhIjoiY2x5ZHRtZDJqMDVsNDJrb3VmZWZoMG9yciJ9.Vid6j50Ey1xMLT6n6g6AgQ
