+++ src/pages/index.tsx
@@ -1,9 +1,12 @@
 import React, { useState, useEffect, useRef } from 'react';
-import { useRouter } from 'next/router';
 import mapboxgl from 'mapbox-gl';
 import 'mapbox-gl/dist/mapbox-gl.css';
+import RegionSelector from '../components/RegionSelector';
+import RouteSelector from '../components/RouteSelector';
+import ServiceSelector from '../components/ServiceSelector';
+import StopsDisplay from '../components/StopsDisplay';
 
-mapboxgl.accessToken = 'pk.eyJ1IjoiemFjYm1yMjIiLCJhIjoiY2x5ZHRtZDJqMDVsNDJrb3VmZWZoMG9yciJ9.Vid6j50Ey1xMLT6n6g6AgQ';
+mapboxgl.accessToken = process.env.NEXT_PUBLIC_MAPBOX_ACCESS_TOKEN || '';
 
 const CombinedPage: React.FC = () => {
   const [selectedArea, setSelectedArea] = useState<string | null>(null);
@@ -18,12 +21,13 @@ const CombinedPage: React.FC = () => {
   const [loading, setLoading] = useState<boolean>(true);
   const [visibleColumn, setVisibleColumn] = useState<number>(0);
 
-  const router = useRouter();
+  const API_BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
+
 
   useEffect(() => {
     const fetchRegions = async () => {
       try {
-        const response = await fetch('https://bus-app-api-kl95.onrender.com/region_data_app');
+        const response = await fetch(`${API_BASE_URL}/region_data_app`);
         const data = await response.json();
         setRegions(data.data);
         setLoading(false);
@@ -55,7 +59,7 @@ const CombinedPage: React.FC = () => {
 
   const fetchTimetableData = async (region: string) => {
     try {
-      const response = await fetch(`https://bus-app-api-kl95.onrender.com/timetable_data_app/${region}`);
+      const response = await fetch(`${API_BASE_URL}/timetable_data_app/${region}`);
       const data = await response.json();
       setTimetableData({ [region]: data.routes });
     } catch (error) {
@@ -104,16 +108,6 @@ const CombinedPage: React.FC = () => {
     return versionData ? versionData.stops : [];
   };
 
-  const handleNextColumn = () => {
-    setVisibleColumn((prevColumn) => prevColumn + 1);
-  };
-
-  const handlePrevColumn = () => {
-    if (visibleColumn > 0) {
-      setVisibleColumn((prevColumn) => prevColumn - 1);
-    }
-  };
-
   // Fixing the time increment feature
   const calculateStopTime = (startTime: string, increment: number) => {
     const [hours, minutes] = startTime.split(':').map(Number);
@@ -140,139 +134,37 @@ const CombinedPage: React.FC = () => {
           {loading ? (
             <p>Loading regions...</p>
           ) : currentPage === 1 ? (
-            <div className="text-center">
-              <h2 className="text-3xl font-semibold mb-6">Choose Your Region</h2>
-              <div className="grid grid-cols-1 sm:grid-cols-2 gap-6">
-                {regions.length > 0 ? (
-                  regions.map((region: any) => (
-                    <button
-                      key={region.id}
-                      className="m-2 p-4 font-bold rounded-lg shadow-lg transform transition-transform duration-300 hover:scale-105"
-                      onClick={() => handleAreaSelect(region.id)}
-                      style={{ backgroundColor: '#FFFACD', color: 'black' }}
-                    >
-                      {region.region_name}
-                    </button>
-                  ))
-                ) : (
-                  <p>No regions available.</p>
-                )}
-              </div>
-            </div>
+            <RegionSelector regions={regions} onAreaSelect={handleAreaSelect} />
           ) : null}
 
           {currentPage === 2 && selectedArea && (
-            <div className="text-center">
-              <h2 className="text-3xl font-semibold mb-6">Select a Route</h2>
-              {timetableData[selectedArea] && timetableData[selectedArea].length > 0 ? (
-                <div className="grid grid-cols-1 sm:grid-cols-3 gap-6">
-                  {timetableData[selectedArea].map((route: any) => (
-                    <button
-                      key={route.title}
-                      className="m-2 p-4 font-bold rounded-lg shadow-lg transform transition-transform duration-300 hover:scale-105"
-                      onClick={() => handleRouteSelect(route)}
-                      style={{ backgroundColor: '#FFFACD', color: 'black' }}
-                    >
-                      {route.title}
-                    </button>
-                  ))}
-                </div>
-              ) : (
-                <p>No routes available for this region.</p>
-              )}
-              <button
-                className="mt-6 px-4 py-2 bg-gradient-to-r from-red-500 to-pink-500 text-white font-bold rounded-lg shadow-lg hover:from-red-600 hover:to-pink-600 transform transition-transform duration-300 hover:scale-105"
-                onClick={goBack}
-              >
-                Back to Regions
-              </button>
-            </div>
+            <RouteSelector
+              selectedArea={selectedArea}
+              timetableData={timetableData}
+              onRouteSelect={handleRouteSelect}
+              onBack={goBack}
+            />
           )}
 
           {currentPage === 3 && selectedRoute && (
-            <div className="text-center">
-              <h2 className="text-3xl font-semibold mb-6">Services for Route {selectedRoute.title}</h2>
-              <div className="grid grid-cols-1 gap-6">
-                {selectedRoute.services && selectedRoute.services.length > 0 ? (
-                  selectedRoute.services.map((service: any) => (
-                    <div
-                      key={service.code}
-                      className="p-4 bg-blue-100 rounded-lg shadow-lg transform transition-transform duration-300 hover:scale-105 cursor-pointer"
-                      onClick={() => {
-                        setSelectedService(service);
-                        setCurrentPage(4);
-                      }}
-                    >
-                      <h3 className="text-lg font-bold text-blue-700 mb-1">Service {service.code}</h3>
-                      <p className="text-sm text-gray-700">{service.direction}</p>
-                    </div>
-                  ))
-                ) : (
-                  <p>No services available for this route.</p>
-                )}
-              </div>
-              <button
-                className="mt-6 px-4 py-2 bg-gradient-to-r from-red-500 to-pink-500 text-white font-bold rounded-lg shadow-lg hover:from-red-600 hover:to             pink-600 transform transition-transform duration-300 hover:scale-105"
-                onClick={goBack}
-              >
-                Back to Routes
-              </button>
-            </div>
+            <ServiceSelector
+              selectedRoute={selectedRoute}
+              onServiceSelect={(service) => {
+                setSelectedService(service);
+                setCurrentPage(4);
+              }}
+              onBack={goBack}
+            />
           )}
 
-          {/* Stops display */}
           {currentPage === 4 && selectedService && (
-            <div className="w-full max-w-4xl bg-white p-6 rounded-lg shadow-lg mt-8">
-              <h2 className="text-3xl font-semibold mb-6 text-center">
-                Stops for {selectedService.code}
-              </h2>
-
-              <div className="flex justify-between items-center mb-4">
-                <button
-                  className="px-4 py-2 bg-blue-500 text-white font-bold rounded-lg shadow-lg"
-                  onClick={handlePrevColumn}
-                  disabled={visibleColumn === 0}
-                >
-                  &lt;
-                </button>
-                <button
-                  className="px-4 py-2 bg-blue-500 text-white font-bold rounded-lg shadow-lg"
-                  onClick={handleNextColumn}
-                >
-                  &gt;
-                </button>
-              </div>
-
-              {/* Stops Table */}
-              <table className="min-w-full table-auto">
-                <thead className="bg-gray-50">
-                  <tr>
-                    <th className="px-6 py-3 text-left text-sm font-medium text-gray-700">Stop Name</th>
-                    <th className="px-6 py-3 text-left text-sm font-medium text-gray-700">Time</th>
-                  </tr>
-                </thead>
-                <tbody className="bg-white divide-y divide-gray-200">
-                  {getCurrentDayTrips(selectedService).map((trip: any, index: number) => (
-                    <React.Fragment key={index}>
-                      {getStopsForCurrentServiceVersion(selectedService, trip.service_version).map((stop: any, stopIndex: number) => (
-                        <tr key={stopIndex}>
-                          <td className="px-6 py-4 text-sm text-gray-700">{stop.address}</td>
-                          <td className="px-6 py-4 text-sm text-gray-700">
-                            {calculateStopTime(trip.start_time, stop.increment + visibleColumn * 30)}
-                          </td>
-                        </tr>
-                      ))}
-                    </React.Fragment>
-                  ))}
-                </tbody>
-              </table>
-              <button
-                className="mt-6 px-4 py-2 bg-gradient-to-r from-red-500 to-pink-500 text-white font-bold rounded-lg shadow-lg"
-                onClick={goBack}
-              >
-                Back to Services
-              </button>
-            </div>
+            <StopsDisplay
+              selectedService={selectedService}
+              getCurrentDayTrips={getCurrentDayTrips}
+              getStopsForCurrentServiceVersion={getStopsForCurrentServiceVersion}
+              calculateStopTime={calculateStopTime}
+              onBack={goBack}
+            />
           )}
         </div>
       </div>
@@ -288,4 +180,3 @@ const CombinedPage: React.FC = () => {
 };
 
 export default CombinedPage;
-
